Refactored a huge bunch of code: renamed module pkg -> internal, removed postgres implementations, improved repositories' implementations and other
