+++ cmd/main.go
@@ -1,25 +1,15 @@
 package main
 
 import (
-	"chattweiler/pkg/bot"
-	"chattweiler/pkg/configs"
-	"chattweiler/pkg/repository"
-	"chattweiler/pkg/utils"
-
+	"chattweiler/internal/bot"
+	"chattweiler/internal/repository/factory"
 	_ "github.com/lib/pq"
 )
 
 func main() {
-	var createRepositoryType repository.StorageType
-	if utils.GetEnvOrDefault(configs.PgDatasourceString) != "" {
-		createRepositoryType = repository.Postgresql
-	} else {
-		createRepositoryType = repository.CsvYandexObjectStorage
-	}
-
 	bot.NewLongPoolingBot(
-		repository.CreatePhraseRepository(createRepositoryType),
-		repository.CreateMembershipWarningRepository(createRepositoryType),
-		repository.CreateContentSourceRepository(createRepositoryType),
+		factory.CreatePhraseRepository(factory.CsvYandexObjectStorage),
+		factory.CreateMembershipWarningRepository(factory.CsvYandexObjectStorage),
+		factory.CreateContentSourceRepository(factory.CsvYandexObjectStorage),
 	).Serve()
 }
