+++ internal/bot/bot.go
@@ -1,0 +1,17 @@
+package bot
+
+import (
+	"chattweiler/internal/bot/object"
+)
+
+var logPackage = "bot"
+
+type Bot interface {
+	Serve()
+
+	handleChatUserJoinEvent(*object.ChatEvent)
+	handleChatUserLeavingEvent(*object.ChatEvent)
+
+	handleInfoCommand(*object.ChatEvent)
+	handleContentRequestCommand(command *object.ContentRequestCommand)
+}
