+++ internal/configs/static.go
