+++ internal/repository/storage/storage.go
@@ -1,0 +1,3 @@
+package storage
+
+var logPackage = "storage"
