+++ internal/vk/content/content.go
@@ -1,0 +1,9 @@
+package content
+
+import (
+	"github.com/SevereCloud/vksdk/v2/object"
+)
+
+type AttachmentsContentCollector interface {
+	CollectOne() object.WallWallpostAttachment
+}
