--- pkg/utils/math.go
@@ -1,21 +1,0 @@
-package utils
-
-func ClampFloat32(value float32, left float32, right float32) float32 {
-	if value > right {
-		return right
-	} else if value < left {
-		return left
-	}
-
-	return value
-}
-
-func ClampInt(value int, left int, right int) int {
-	if value > right {
-		return right
-	} else if value < left {
-		return left
-	}
-
-	return value
-}
