--- pkg/vk/content/content.go
@@ -1,10 +1,0 @@
-// Package content provides services for content fetching
-package content
-
-import (
-	"github.com/SevereCloud/vksdk/v2/object"
-)
-
-type AttachmentsContentCollector interface {
-	CollectOne() object.WallWallpostAttachment
-}
